refactor(reservation-domain): Reservation 패키지로 도메인 분리
- Reservation, 관련 Enum 대상으로 domain.reservation 패키지로 이동함
- Test, TestFixture 도 패키지 맞춤
