+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/booking/contract/ReservationOwnershipPolicy.java
@@ -2,7 +2,7 @@ package com.seatliberator.seatliberator.reservation.application.booking.contract
 
 import com.seatliberator.seatliberator.identity.core.actor.Actor;
 import com.seatliberator.seatliberator.reservation.application.booking.contract.result.ReservationPolicyResult;
-import com.seatliberator.seatliberator.reservation.domain.persistence.Reservation;
+import com.seatliberator.seatliberator.reservation.domain.reservation.Reservation;
 
 public interface ReservationOwnershipPolicy {
     ReservationPolicyResult evaluate(Reservation reservation, Actor requester);
