+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/booking/port/in/query/FindMyReservationQuery.java
@@ -1,7 +1,7 @@
 package com.seatliberator.seatliberator.reservation.application.booking.port.in.query;
 
-import com.seatliberator.seatliberator.reservation.domain.ReservationStatus;
 import com.seatliberator.seatliberator.reservation.domain.TimeRange;
+import com.seatliberator.seatliberator.reservation.domain.reservation.ReservationStatus;
 
 public record FindMyReservationQuery(
         String userId,
