+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/application/waitlist/handler/SeatVacancyHandler.java
@@ -8,8 +8,8 @@ import com.seatliberator.seatliberator.reservation.application.waitlist.port.in.
 import com.seatliberator.seatliberator.reservation.application.waitlist.port.out.WaitlistStore;
 import com.seatliberator.seatliberator.reservation.domain.SeatLocator;
 import com.seatliberator.seatliberator.reservation.domain.TimeRange;
-import com.seatliberator.seatliberator.reservation.domain.event.ReservationCanceled;
-import com.seatliberator.seatliberator.reservation.domain.event.ReservationExpired;
+import com.seatliberator.seatliberator.reservation.domain.reservation.event.ReservationCanceled;
+import com.seatliberator.seatliberator.reservation.domain.reservation.event.ReservationExpired;
 import com.seatliberator.seatliberator.reservation.domain.waitlist.WaitlistStatus;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
