+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/book/application/model/ReservationOccupancyPolicyTest.java
@@ -1,7 +1,7 @@
 package com.seatliberator.seatliberator.reservation.book.application.model;
 
 import com.seatliberator.seatliberator.reservation.application.booking.model.ReservationOccupancyPolicy;
-import com.seatliberator.seatliberator.reservation.domain.ReservationStatus;
+import com.seatliberator.seatliberator.reservation.domain.reservation.ReservationStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
