+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/waitlist/application/handler/SeatVacancyHandlerTest.java
@@ -7,8 +7,8 @@ import com.seatliberator.seatliberator.reservation.application.waitlist.internal
 import com.seatliberator.seatliberator.reservation.application.waitlist.port.out.WaitlistStore;
 import com.seatliberator.seatliberator.reservation.domain.SimpleSeatLocator;
 import com.seatliberator.seatliberator.reservation.domain.SimpleTimeRange;
-import com.seatliberator.seatliberator.reservation.domain.event.ReservationCanceled;
-import com.seatliberator.seatliberator.reservation.domain.event.ReservationExpired;
+import com.seatliberator.seatliberator.reservation.domain.reservation.event.ReservationCanceled;
+import com.seatliberator.seatliberator.reservation.domain.reservation.event.ReservationExpired;
 import com.seatliberator.seatliberator.reservation.domain.waitlist.Waitlist;
 import com.seatliberator.seatliberator.reservation.domain.waitlist.WaitlistFixtureBuilder;
 import com.seatliberator.seatliberator.reservation.domain.waitlist.WaitlistStatus;
