--- reservation/reservation-domain/src/main/java/com/seatliberator/seatliberator/reservation/domain/event/DomainEvent.java
@@ -1,4 +1,0 @@
-package com.seatliberator.seatliberator.reservation.domain.event;
-
-public interface DomainEvent {
-}
