+++ reservation/reservation-domain/src/main/java/com/seatliberator/seatliberator/reservation/domain/reservation/ReservationStatus.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.reservation.domain;
+package com.seatliberator.seatliberator.reservation.domain.reservation;
 
 public enum ReservationStatus {
     RESERVED,
