+++ reservation/reservation-domain/src/main/java/com/seatliberator/seatliberator/reservation/domain/reservation/event/DomainEvent.java
@@ -1,0 +1,4 @@
+package com.seatliberator.seatliberator.reservation.domain.reservation.event;
+
+public interface DomainEvent {
+}
