+++ reservation/reservation-domain/src/main/java/com/seatliberator/seatliberator/reservation/domain/reservation/event/ReservationCreated.java
@@ -1,13 +1,10 @@
-package com.seatliberator.seatliberator.reservation.domain.event;
+package com.seatliberator.seatliberator.reservation.domain.reservation.event;
 
 import com.seatliberator.seatliberator.reservation.domain.SeatLocator;
 import com.seatliberator.seatliberator.reservation.domain.TimeRange;
 
-import java.time.Instant;
-
-public record ReservationExpired(
+public record ReservationCreated(
         SeatLocator locator,
-        TimeRange range,
-        Instant detectedAt
+        TimeRange range
 ) implements DomainEvent {
 }
