+++ reservation/reservation-domain/src/main/java/com/seatliberator/seatliberator/reservation/domain/reservation/event/ReservationExpired.java
@@ -1,13 +1,13 @@
-package com.seatliberator.seatliberator.reservation.domain.event;
+package com.seatliberator.seatliberator.reservation.domain.reservation.event;
 
 import com.seatliberator.seatliberator.reservation.domain.SeatLocator;
 import com.seatliberator.seatliberator.reservation.domain.TimeRange;
 
 import java.time.Instant;
 
-public record ReservationCanceled(
+public record ReservationExpired(
         SeatLocator locator,
         TimeRange range,
-        Instant canceledAt
+        Instant detectedAt
 ) implements DomainEvent {
 }
