+++ reservation/reservation-persistence/src/main/java/com/seatliberator/seatliberator/reservation/persistence/book/jpa/repository/ReservationRepository.java
@@ -1,6 +1,6 @@
 package com.seatliberator.seatliberator.reservation.persistence.book.jpa.repository;
 
-import com.seatliberator.seatliberator.reservation.domain.persistence.Reservation;
+import com.seatliberator.seatliberator.reservation.domain.reservation.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
